refact: extrair combat dialogs para module/combat/combat-dialogs.mjs
